+++ uint64_slice.c
@@ -208,7 +208,7 @@ uint64_slice_sort(uint64_slice_t *s, sort_compare_func_t sort_compare)
 		return ;
 	}
 
-	qsort(s->items, s->len, sizeof(int), sort_compare);
+	qsort(s->items, s->len, sizeof(uint64_t), sort_compare);
 }
 
 uint64_t
